refactor(ingress-egress-tracker): give structure
